% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{fourth_root_power_trans}
\alias{fourth_root_power_trans}
\title{Fourth root scale transformation for ggplot}
\usage{
fourth_root_power_trans()
}
\description{
Fourth root scale transformation for ggplot
}
\examples{
ggplot2::ggplot(mtcars) +
  ggplot2::geom_point(ggplot2::aes(x = wt, colour = mpg, y = factor(gear))) +
  ggplot2::scale_colour_viridis_c(trans = "fourth_root_power")
}
\references{
\url{https://andrewgelman.com/2007/01/02/the_14power_tra/}
}
