% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sleek.R
\name{theme_sleek}
\alias{theme_sleek}
\title{A sleek theme for ggplot2}
\usage{
theme_sleek(base_size = 11, base_family = "")
}
\description{
A simple theme for ggplot2 that loosely resembles nicely themed plots from
base graphics.
}
\examples{
p <- ggplot2::ggplot(mtcars) +
  ggplot2::geom_point(ggplot2::aes(x = wt, y = mpg, colour = factor(gear))) +
  ggplot2::facet_wrap(~am)
p + theme_sleek()

}

